Feat: Refactor api
