+++ ataraxai/routes/api_message.py
@@ -1,0 +1,10 @@
+
+
+
+
+class Messages:
+    VAULT_LOCKED = "Vault is locked"
+    VAULT_UNLOCKED = "Vault is already unlocked"
+    PROJECT_NOT_FOUND = "Project not found"
+    SESSION_NOT_FOUND = "Session not found"
+    OPERATION_FAILED = "An error occurred while processing the request"
