+++ ataraxai/routes/dependency_api.py
@@ -1,0 +1,32 @@
+from typing import Any
+from fastapi import HTTPException, Request
+from fastapi.params import Depends
+from ataraxai.praxis.ataraxai_orchestrator import AtaraxAIOrchestrator
+from fastapi import status
+from ataraxai.praxis.utils.app_state import AppState
+from ataraxai.routes.api_message import Messages
+
+def get_orchestrator(request: Request) -> AtaraxAIOrchestrator:
+    return request.app.state.orchestrator
+
+
+def get_unlocked_orchestrator(
+    orch: Any = Depends(get_orchestrator),
+) -> AtaraxAIOrchestrator:
+    if orch.state != AppState.UNLOCKED:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN, detail=Messages.VAULT_LOCKED
+        )
+    return orch
+
+def katalepsis_monitor(request: Request) -> Any:
+    """
+    Dependency to get the Katalepsis instance from the FastAPI application state.
+    
+    Args:
+        request (Request): The FastAPI request object.
+
+    Returns:
+        Katalepsis: The Katalepsis instance from the application state.
+    """
+    return request.app.state.katalepsis
